// src/types/api/common.d.ts
export type Category =
  | "backgrounds"
  | "fashion"
  | "nature"
  | "science"
  | "education"
  | "feelings"
  | "health"
  | "people"
  | "religion"
  | "places"
  | "animals"
  | "industry"
  | "computer"
  | "food"
  | "sports"
  | "transportation"
  | "travel"
  | "buildings"
  | "business"
  | "music";

export type ImageType = "all" | "photo" | "illustration" | "vector";
export type VideoType = "all" | "film" | "animation";
export type OrderType = "popular" | "latest";